backend/requirements.txt
# Multimodal Radiological AI Assistant - Backend Dependencies

# Core Framework
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Deep Learning & Vision
torch>=2.0.0
torchvision>=0.15.0
torchxrayvision>=1.2.0

# NLP & Embeddings
sentence-transformers>=2.2.0

# LLM - Google Gemini API
google-generativeai>=0.3.0

# RAG & Search - LlamaIndex
llama-index>=0.10.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-embeddings-huggingface>=0.2.0
faiss-cpu>=1.7.4

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
scikit-image>=0.24.0
Pillow>=10.1.0

# Dataset
kagglehub>=0.3.4  # Latest version with dataset_download support

# Utilities
python-dotenv==1.0.0
